+++ init.h
@@ -187,7 +187,7 @@ int searchTudoClient(Tudo tudo, char* cliente);
 
 double getFatMesTudo(const Tudo tudo, const char* produto, Tipo tipo, Filial filial, int mes);
 
-double getQuantMesTudo(const Tudo tudo, const char* produto, Tipo tipo, Filial filial, int mes);
+int getQuantMesTudo(const Tudo tudo, const char* produto, Tipo tipo, Filial filial, int mes);
 
 /**
 @brief Libertar um inicializador
